Renamed buildMocks to buildTest
